requirements.txt
Django>=4.2.6
django-allauth>=0.58
psycopg>=3.1.12
psycopg-binary>=3.1.12
dj-rest-auth>=5.0
python-decouple>=3.8
djangorestframework>=3.14
markdown>=3.5
django-filter>=23.3
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
google_api_python_client>=2.1
google_auth_oauthlib>=1.1
google-auth-httplib2>=0.1
django-storages[s3]>=1.14
Pillow>=10.1
drf-spectacular>=0.26
python-dateutil>=2.8